move public key, secret key and signature struct holder header files under `internals` directory
