+++ tests/test_signing.cpp
@@ -1,5 +1,5 @@
-#include "raccoon/public_key.hpp"
-#include "raccoon/secret_key.hpp"
+#include "raccoon/internals/public_key.hpp"
+#include "raccoon/internals/secret_key.hpp"
 #include <gtest/gtest.h>
 
 // Test key generation -> signing -> verification flow for random message of given byte length.
